backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-CORS==4.0.0
Flask-WTF==1.2.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
WTForms==3.1.1
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
email-validator==2.1.0
